requirements.txt
# Core framework
fastapi
uvicorn[standard]

# Data validation & settings
pydantic
pydantic-settings

# Database
supabase
postgrest

# Resilience
tenacity
# Auth / JWT
python-jose[cryptography]

# LLM vendors (install only those you use)
anthropic
openai
google-generativeai
groq

# Payments
stripe

# PDF generation
weasyprint

# File uploads in FastAPI
python-multipart

# Email
aiosmtplib